Make DiffGraphApplier public again.
